crud_api/requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
aws-cdk.assets==1.116.0
aws-cdk.aws-apigateway==1.115.0
aws-cdk.aws-apigatewayv2==1.115.0
aws-cdk.aws-apigatewayv2-integrations==1.115.0
aws-cdk.aws-applicationautoscaling==1.116.0
aws-cdk.aws-autoscaling-common==1.116.0
aws-cdk.aws-certificatemanager==1.115.0
aws-cdk.aws-cloudformation==1.115.0
aws-cdk.aws-cloudwatch==1.116.0
aws-cdk.aws-codeguruprofiler==1.116.0
aws-cdk.aws-codestarnotifications==1.115.0
aws-cdk.aws-cognito==1.115.0
aws-cdk.aws-dynamodb==1.115.0
aws-cdk.aws-ec2==1.116.0
aws-cdk.aws-ecr==1.116.0
aws-cdk.aws-ecr-assets==1.116.0
aws-cdk.aws-efs==1.116.0
aws-cdk.aws-elasticloadbalancingv2==1.115.0
aws-cdk.aws-events==1.116.0
aws-cdk.aws-iam==1.116.0
aws-cdk.aws-kinesis==1.115.0
aws-cdk.aws-kms==1.116.0
aws-cdk.aws-lambda==1.116.0
aws-cdk.aws-lambda-python==1.116.0
aws-cdk.aws-logs==1.116.0
aws-cdk.aws-rds==1.115.0
aws-cdk.aws-route53==1.115.0
aws-cdk.aws-s3==1.116.0
aws-cdk.aws-s3-assets==1.116.0
aws-cdk.aws-sam==1.115.0
aws-cdk.aws-secretsmanager==1.115.0
aws-cdk.aws-servicediscovery==1.115.0
aws-cdk.aws-signer==1.116.0
aws-cdk.aws-sns==1.115.0
aws-cdk.aws-sqs==1.116.0
aws-cdk.aws-ssm==1.116.0
aws-cdk.cloud-assembly-schema==1.116.0
aws-cdk.core==1.116.0
aws-cdk.custom-resources==1.115.0
aws-cdk.cx-api==1.116.0
aws-cdk.region-info==1.116.0
backcall==0.2.0
bleach==3.3.1
boto3==1.18.12
botocore==1.21.12
cattrs==1.5.0
cffi==1.14.6
constructs==3.3.98
# Editable Git install with no remote (crud-api==0.0.1)
-e /Users/ss/projects/proof-of-concept/crud_api/crud_api
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsii==1.31.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mysql-connector==2.2.9
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
publication==0.0.3
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
pyzmq==22.1.0
s3transfer==0.5.0
Send2Trash==1.7.1
simplejson==3.17.3
six==1.16.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
